Moved src/types.rs tests into relevant files
