+++ src/types.rs
@@ -137,124 +137,3 @@ impl CacheOptions {
         self.cache_immutable_data
     }
 }
-
-#[cfg(test)]
-#[allow(deprecated)]
-mod test {
-    extern crate cbor;
-    use super::*;
-    use sodiumoxide::crypto;
-    use std::cmp;
-    use rustc_serialize::{Decodable, Encodable};
-    use test_utils::Random;
-    use public_id::PublicId;
-    use authority::Authority;
-    use NameType;
-    use name_type::closer_to_target;
-    use sodiumoxide::crypto::sign;
-    use utils;
-
-    fn test_object<T>(obj_before : T) where T: for<'a> Encodable + Decodable + Eq {
-        let mut e = cbor::Encoder::from_memory();
-        e.encode(&[&obj_before]).unwrap();
-        let mut d = cbor::Decoder::from_bytes(e.as_bytes());
-        let obj_after: T = d.decode().next().unwrap().unwrap();
-        assert_eq!(obj_after == obj_before, true)
-
-    }
-
-    #[test]
-    fn test_authority() {
-        test_object(Authority::ClientManager(Random::generate_random()));
-        test_object(Authority::NaeManager(Random::generate_random()));
-        test_object(Authority::NodeManager(Random::generate_random()));
-        test_object(Authority::ManagedNode(Random::generate_random()));
-        test_object(Authority::Client(Random::generate_random(), sign::gen_keypair().0));
-    }
-
-    #[test]
-    fn serialisation_public_id() {
-        let obj_before = PublicId::generate_random();
-
-        let mut e = cbor::Encoder::from_memory();
-        e.encode(&[&obj_before]).unwrap();
-
-        let mut d = cbor::Decoder::from_bytes(e.as_bytes());
-        let obj_after: PublicId = d.decode().next().unwrap().unwrap();
-        assert_eq!(obj_before, obj_after);
-    }
-
-    #[test]
-    fn test_calculate_relocated_name() {
-        let original_name : NameType = Random::generate_random();
-
-        // empty close nodes
-        assert!(utils::calculate_relocated_name(Vec::new(), &original_name).is_err());
-
-        // one entry
-        let mut close_nodes_one_entry : Vec<NameType> = Vec::new();
-        close_nodes_one_entry.push(Random::generate_random());
-        let actual_relocated_name_one_entry = utils::calculate_relocated_name(close_nodes_one_entry.clone(),
-        &original_name).unwrap();
-        assert!(original_name != actual_relocated_name_one_entry);
-
-        let mut combined_one_node_vec : Vec<NameType> = Vec::new();
-        combined_one_node_vec.push(original_name.clone());
-        combined_one_node_vec.push(close_nodes_one_entry[0].clone());
-
-        let mut combined_one_node: Vec<u8> = Vec::new();
-        for node_id in combined_one_node_vec {
-            for i in node_id.get_id().iter() {
-                combined_one_node.push(*i);
-            }
-        }
-
-        let expected_relocated_name_one_node =
-            NameType(crypto::hash::sha512::hash(&combined_one_node).0);
-
-        assert_eq!(actual_relocated_name_one_entry, expected_relocated_name_one_node);
-
-        // populated closed nodes
-        let mut close_nodes : Vec<NameType> = Vec::new();
-        for _ in 0..GROUP_SIZE {
-            close_nodes.push(Random::generate_random());
-        }
-        let actual_relocated_name = utils::calculate_relocated_name(close_nodes.clone(),
-        &original_name).unwrap();
-        assert!(original_name != actual_relocated_name);
-        close_nodes.sort_by(|a, b| if closer_to_target(&a, &b, &original_name) {
-            cmp::Ordering::Less
-        } else {
-            cmp::Ordering::Greater
-        });
-        let first_closest = close_nodes[0].clone();
-        let second_closest = close_nodes[1].clone();
-        let mut combined: Vec<u8> = Vec::new();
-
-        for i in original_name.get_id().into_iter() {
-            combined.push(*i);
-        }
-        for i in first_closest.get_id().into_iter() {
-            combined.push(*i);
-        }
-        for i in second_closest.get_id().into_iter() {
-            combined.push(*i);
-        }
-
-        let expected_relocated_name = NameType(crypto::hash::sha512::hash(&combined).0);
-        assert_eq!(expected_relocated_name, actual_relocated_name);
-
-        let mut invalid_combined: Vec<u8> = Vec::new();
-        for i in first_closest.get_id().into_iter() {
-            invalid_combined.push(*i);
-        }
-        for i in second_closest.get_id().into_iter() {
-            invalid_combined.push(*i);
-        }
-        for i in original_name.get_id().into_iter() {
-            invalid_combined.push(*i);
-        }
-        let invalid_relocated_name = NameType(crypto::hash::sha512::hash(&invalid_combined).0);
-        assert!(invalid_relocated_name != actual_relocated_name);
-    }
-}
